Добавлено логирование запросов и ответов в сервис gateway

+++ transferService/src/main/java/ru/muravin/bankapplication/transferService/dto/CheckedTransferDto.java
@@ -6,6 +6,6 @@ import lombok.Data;
 public class CheckedTransferDto {
     private String fromAccountNumber;
     private String toAccountNumber;
-    private Float amountFrom;
-    private Float amountTo;
+    private Double amountFrom;
+    private Double amountTo;
 }
